requirements.txt
pandas
requests
discord.py
beautifulsoup4
discord
python-dotenv
html5lib
emojis